v5.1 mimik Client Library Trial
v5.1 mimik Client Library Trial

The purpose of the mimik Client Library Trial is to provide a programmatic way to work with the edgeEngine Runtime to access information about the mobile device on which the application is running, as well as mobile devices running within a cluster of mobile devices that are hosting the edgeEngine Runtime. Also, to allow developers to use edge microservices running within a particular cluster.